Converting code to Groovy
