+++ .gitignore
@@ -1,3 +1,5 @@
 .idea/
 *.iml
 target/
+build/
+.gradle/
