+++ src/main/groovy/quantum/dao/StockNonBlockingDatasource.groovy
@@ -1,0 +1,11 @@
+package quantum.dao
+
+import quantum.domain.Movimento
+import rx.Observable
+
+/**
+ * Created by xonda on 23/03/2015.
+ */
+public interface StockNonBlockingDatasource {
+    Observable<Movimento> list()
+}
