--- src/main/java/br/org/quantum/domain/VolumeMedio.java
@@ -1,25 +1,0 @@
-package br.org.quantum.domain;
-
-/**
- * Created by xonda on 15/03/2015.
- */
-public class VolumeMedio {
-    private final String id;
-    private final Double volume;
-
-
-
-    public VolumeMedio(String id, Double volume) {
-        this.id = id;
-        this.volume = volume;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public Double getVolume() {
-        return volume;
-    }
-
-}
