--- src/main/java/br/org/quantum/mvc/HomeController.java
@@ -1,11 +1,0 @@
-package br.org.quantum.mvc;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class HomeController {
-    @RequestMapping("/home")
-    public String home() {
-        return "index";
-    }
-}
