--- src/main/resources/public/table-volume.html
@@ -1,10 +1,0 @@
-<table class="table table-striped">
-    <tr>
-        <th>Ação</th>
-        <th>Volume</th>
-    </tr>
-    <tr ng-repeat="volume in volumes">
-        <td>{{volume.id}}</td>
-        <td>{{volume.volume}}</td>
-    </tr>
-</table>
