--- src/main/resources/static/style/style.css
@@ -1,7 +1,0 @@
-body {
-    padding-top: 50px;
-}
-.starter-template {
-    padding: 40px 15px;
-    text-align: center;
-}
